# NumDigits: minor speedups in lookup table logic