--- a/table.go
+++ b/table.go
@@ -25,13 +25,17 @@ import "math/big"
 // Using this proof, for a given digit count, the map will return the lower number
 // of decimal digits (k) the binary digit count could represent, along with the
 // value of the border between the two decimal digit counts (10^k).
-const digitsTableSize = 128
+const (
+	digitsTableSize = 128
+	digitsTableLen  = digitsTableSize + 1
+)
 
-var digitsLookupTable [digitsTableSize + 1]tableVal
+var digitsLookupTable [digitsTableLen]tableVal
 
 type tableVal struct {
-	digits int64
-	border big.Int
+	digits  int64
+	border  big.Int
+	nborder big.Int
 }
 
 func init() {
@@ -48,14 +52,8 @@ func init() {
 		elem.border.SetInt64(10)
 		curExp.SetInt64(elem.digits)
 		elem.border.Exp(&elem.border, curExp, nil)
+		elem.nborder.Neg(&elem.border)
 	}
-}
-
-func lookupBits(bitLen int) (tableVal, bool) {
-	if bitLen > 0 && bitLen < len(digitsLookupTable) {
-		return digitsLookupTable[bitLen], true
-	}
-	return tableVal{}, false
 }
 
 // NumDigits returns the number of decimal digits of d.Coeff.
@@ -69,17 +67,30 @@ func NumDigits(b *big.Int) int64 {
 	if bl == 0 {
 		return 1
 	}
-	if val, ok := lookupBits(bl); ok {
-		ab := new(big.Int).Abs(b)
-		if ab.Cmp(&val.border) < 0 {
-			return val.digits
+
+	if bl < digitsTableLen {
+		val := digitsLookupTable[bl]
+		switch b.Sign() {
+		case 1:
+			if b.Cmp(&val.border) < 0 {
+				return val.digits
+			}
+		case -1:
+			if b.Cmp(&val.nborder) > 0 {
+				return val.digits
+			}
 		}
 		return val.digits + 1
 	}
 
 	n := int64(float64(bl) / digitsToBitsRatio)
-	a := new(big.Int).Abs(b)
-	e := new(big.Int).Exp(bigTen, big.NewInt(n), nil)
+	a := new(big.Int)
+	e := tableExp10(n, a)
+	if b.Sign() < 0 {
+		a.Abs(b)
+	} else {
+		a = b
+	}
 	if a.Cmp(e) >= 0 {
 		n++
 	}
@@ -90,7 +101,7 @@ func NumDigits(b *big.Int) int64 {
 // is stored in the pow10LookupTable. For instance, if the powerTenTableSize
 // if 3, then the lookup table will store power of 10 values from 10^0 to
 // 10^3 inclusive.
-const powerTenTableSize = 64
+const powerTenTableSize = 128
 
 var pow10LookupTable [powerTenTableSize + 1]big.Int