Also don't take Abs when not needed.

```
name                old time/op    new time/op    delta
NumDigitsLookup-12     775µs ±14%     381µs ± 3%  -50.80%  (p=0.008 n=5+5)
NumDigitsFull-12      13.0ms ±19%    12.8ms ±11%     ~     (p=1.000 n=5+5)

name                old alloc/op   new alloc/op   delta
NumDigitsLookup-12     115kB ± 0%      22kB ± 0%     ~     (p=0.079 n=4+5)
NumDigitsFull-12      4.49MB ± 0%    3.91MB ± 0%  -12.88%  (p=0.008 n=5+5)

name                old allocs/op  new allocs/op  delta
NumDigitsLookup-12     2.14k ± 0%     0.42k ± 0%  -80.43%  (p=0.008 n=5+5)
NumDigitsFull-12       34.6k ± 0%     31.1k ± 0%  -10.06%  (p=0.008 n=5+5)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/apd/21)
<!-- Reviewable:end -->