@nvanbenschoten 2017-02-17
LGTM, although I'd still like to clear up the `tableExp10` question.