@nvanbenschoten 2017-02-17
Nice speedups, by the way!